feat(auth): Add auth with OAuth2 standard from fastapi.security + GET /users/me + POST /users
Auth respect OAuth2 standard from fastapi.security. I follow and adapt this https://fastapi.tiangolo.com/tutorial/security/oauth2-jwt/
It allow auth with an access_token.  But with a web interface, it seems preferable to add a refresh token in the next version to improve the UX and avoid logging in each time the page is refreshed.
The refresh token needs to be stored in an http-only cookie (because the front-end is a web browser) to enable good security.  And the access token will be store in the web application state.